+++ packages/block-library/src/navigation/edit/index.js
@@ -292,7 +292,7 @@ function Navigation( {
 		// If there's non fallback navigation menus and
 		// a classic menu with a `primary` location or slug,
 		// then create a new navigation menu based on it.
-		// Otherwise, use the first classic menu.
+		// Otherwise, use the most recently created classic menu.
 		const primaryMenus = classicMenus.filter(
 			( classicMenu ) =>
 				classicMenu.locations.includes( 'primary' ) ||
@@ -300,22 +300,15 @@ function Navigation( {
 		);
 
 		if ( primaryMenus.length ) {
-			// Sort by location to allow the menu assigned to primary location to take precedence.
-			primaryMenus.sort( ( a, b ) => {
-				if ( a.locations < b.locations ) {
-					return 1;
-				}
-				if ( a.locations > b.locations ) {
-					return -1;
-				}
-				return 0;
-			} );
 			convertClassicMenu(
 				primaryMenus[ 0 ].id,
 				primaryMenus[ 0 ].name,
 				'publish'
 			);
 		} else {
+			classicMenus.sort( ( a, b ) => {
+				return b.id - a.id;
+			} );
 			convertClassicMenu(
 				classicMenus[ 0 ].id,
 				classicMenus[ 0 ].name,