+++ packages/block-library/src/navigation/index.php
@@ -277,6 +277,12 @@ function block_core_navigation_get_classic_menu_fallback() {
 		}
 
 		// Otherwise return the most recently created classic menu.
+		usort(
+			$classic_nav_menus,
+			function( $a, $b ) {
+				return $b->term_id - $a->term_id;
+			}
+		);
 		return $classic_nav_menus[0];
 	}
 }